// src/app/models/VaccineCenter.model.ts
export class VaccineCenter {
  center_id: number;
  name: string;
  name1?: string;
  address?: string;
  address1?: string;
  state_name: string;
  state_name1?: string;
  district_name: string;
  district_name1?: string;
  block_name: string;
  block_name1?: string;
  pincode: string;
  lat?: number;
  long?: number;
  from: string;
  to: string;
  fee_type: string;
  fee: string;
  session_id: string;
  date: string;
  available_capacity: number;
  available_capacity_dose1?: number;
  available_capacity_dose2?: number;
  min_age_limit: number;
  vaccine: string;
  slots: string[]
}